+++ packages/plantae/CHANGELOG.md
@@ -1,5 +1,17 @@
 # plantae
 
+## 0.2.2
+
+### Patch Changes
+
+- a96136e: fix: settle middleware response for onRejected
+
+## 0.2.1
+
+### Patch Changes
+
+- dbe98be: fix: transform body into null when having null body status
+
 ## 0.2.0
 
 ### Minor Changes